Fix amount not being multiplied in store

+++ LemonadeStand/Item/Cup.cs
@@ -8,7 +8,7 @@ namespace LemonadeStand
 {
     class Cup : Item
     {
-        public Cup(double cost)
+        public Cup(decimal cost)
         {
             this.cost = cost;
         }
